Throw together a way to retrieve all my Amazon books
Apparently Amazon is removing the right to download and archive your
AZW files. This means that you are permanently locked into the Amazon
ecosystem for your books; which is kind of a bummer.

This attempts to idempotently download the files, so if it crashes out
it will not download duplicates.

Run `ruby main.rb --help` for instructions.

Tested on Ruby 3.1.